Match types of counter variable and comparison bound
- TPML_PCR_SELECTION.count is UINT32, so the counter variable needs to
  be UINT32 as well.
- TPMS_PCR_SELECTION.sizeofSelect is UINT8, but after multiplication
  with 8u it might only fit into an unsigned int (which is guaranteed to
  be at least 16 bits long) due to integer promotion.

+++ lib/pcr.c
@@ -191,8 +191,8 @@ bool pcr_print_pcr_struct(TPML_PCR_SELECTION *pcr_select, tpm2_pcrs *pcrs) {
         tpm2_tool_output("  %s:\n", alg_name);
 
         // Loop through all PCRs in this bank
-        UINT8 pcr_id;
-        for (pcr_id = 0; pcr_id < pcr_select->pcrSelections[i].sizeofSelect * 8;
+        unsigned int pcr_id;
+        for (pcr_id = 0; pcr_id < pcr_select->pcrSelections[i].sizeofSelect * 8u;
                 pcr_id++) {
             if (!tpm2_util_is_pcr_select_bit_set(&pcr_select->pcrSelections[i],
                     pcr_id)) {
